Fixed Auth screens ui and added skeleton loading in skill management component
