+++ src/components/Accounts/Accounts.css
@@ -35,13 +35,21 @@
   margin: 2.25rem 4.688rem 1.25rem 4.688rem;
 }
 
-.input-login {
+/* .input-login {
   border: none;
   border-bottom: 0.063rem solid #b2b2b3;
   width: 25.438rem;
   font-size: 0.938rem;
   line-height: 2.563rem;
   max-width: 100%;
+} */
+
+.input-login {
+  border: none;
+  width: 428.58px;
+  font-size: 0.938rem;
+  line-height: 2.563rem;
+  /* max-width: 100%; */
 }
 
 .input-login:focus {
@@ -174,6 +182,13 @@ a:hover {
 
 
 .password-toggle-button {
-  position: absolute;
-    right: 32px;
+  background: none;
+}
+.pass-input-div{
+  display: flex;
+  justify-content: space-between;
+}
+.div-input {
+  border-bottom: 0.063rem solid #b2b2b3;
+  width: 428.58px;
 }
