+++ src/components/UserPortal/SkillManagement/SkillsAdded/SkillsAdded.js
@@ -1,5 +1,5 @@
-import React, { useEffect,useState} from "react";
-import { EmptySkillsAdded } from "./EmptySkillsAdded/EmptySkillsAdded";
+import React, { useEffect, useState } from "react";
+import EmptySkillsAdded from "./EmptySkillsAdded/EmptySkillsAdded";
 import "./SkillsAdded.css";
 // import nonActiveimageStar from '../Assets/nonActiveimageStar.png';
 import { ReactComponent as EmptyStar } from "../../../../Assets/emptystar.svg";
@@ -27,7 +27,7 @@ const SkillsAdded = () => {
       );
       const data = await response.json();
 
-      setAllData(data.response);
+      // setAllData(data.response);
     } catch (error) {
       console.log(error);
     }
@@ -42,15 +42,13 @@ const SkillsAdded = () => {
       {/* //main card  */}
       <div>
         {allData.length === 0 ? (
-          <div>
-            <EmptySkillsAdded />
-          </div>
+          <EmptySkillsAdded/>
         ) : (
           allData.map((DataUsed) => (
             <div className="card" style={{ width: "288px" }}>
               <div
                 class="card-body p-0"
-                style={{ maxHeight: "602px", overflow: "auto" }}
+                style={{ maxHeight:"calc(100vh - 30vh)", overflow: "auto" }}
               >
                 <div className="row cards">
                   <div className="col-12 d-flex mainImg">
